Adds an optimization for overflow-checking subtraction.
When presented with a code snippit like so:

int a, b;
if (a - b == 0) {
  return a + b;
}
return 0;

We need a check for the first `a - b`, but if we fall into the true
case of the if statement, then we know that `b` must be zero. So,
in the body of the true block, we can replace all uses of `b` with
zero. This allows other stages of the compiler to eliminate the second
overflow check for `a + b`.